笔记页面

+++ infinite-api/src/main/java/net/binker/entity/model/TTreDetail.java
@@ -65,6 +65,6 @@ public class TTreDetail {
 
 	//@Comment("文章类型")
 	@Column(name = "type")
-	private Long type;
+	private Integer type;
 
 }
